rename package name Data to Signal
