+++ MACS3/Commands/callpeak_cmd.py
@@ -1,4 +1,4 @@
-# Time-stamp: <2020-11-24 16:44:31 Tao Liu>
+# Time-stamp: <2020-11-25 01:18:46 Tao Liu>
 
 """Description: MACS 2 main executable
 
@@ -22,9 +22,9 @@ import tempfile
 # ------------------------------------
 from MACS3.Utilities.Constants import *
 from MACS3.Utilities.OptValidator import opt_validate_callpeak
-from MACS3.Data.Prob import binomial_cdf_inv
-from MACS3.Data.PeakModel import PeakModel,NotEnoughPairsException
-from MACS3.Data.PeakDetect import PeakDetect
+from MACS3.Signal.Prob import binomial_cdf_inv
+from MACS3.Signal.PeakModel import PeakModel,NotEnoughPairsException
+from MACS3.Signal.PeakDetect import PeakDetect
 from MACS3.IO.OutputWriter import model2r_script
 # ------------------------------------
 # Main function
@@ -257,9 +257,7 @@ def run( args ):
     ofhd_xls = open( options.peakxls, "w" )
     ofhd_xls.write("# This file is generated by MACS version %s\n" % (MACS_VERSION))
     ofhd_xls.write(options.argtxt+"\n")
-
     ofhd_xls.write(tagsinfo)
-
     if options.shift > 0:
         ofhd_xls.write("# Sequencing ends will be shifted towards 3' by %d bp(s)\n" % (options.shift))
     elif options.shift < 0:
@@ -276,6 +274,7 @@ def run( args ):
     # pass write method so we can print too, and include name
     peakdetect.peaks.write_to_xls(ofhd_xls, name = options.name.encode())
     ofhd_xls.close()
+
     #4.2 peaks in BED
     if options.log_pvalue != None:
         score_column = "pscore"
