+++ MACS3/Commands/filterdup_cmd.py
@@ -20,7 +20,7 @@ import logging
 # ------------------------------------
 from MACS3.Utilities.Constants import *
 from MACS3.Utilities.OptValidator import opt_validate_filterdup
-from MACS3.Data.Prob import binomial_cdf_inv
+from MACS3.Signal.Prob import binomial_cdf_inv
 # ------------------------------------
 # Main function
 # ------------------------------------
