+++ MACS3/Signal/PeakModel.pyx
@@ -18,7 +18,7 @@ import array
 # MACS3 modules
 # ------------------------------------
 from MACS3.Utilities.Constants import *
-from MACS3.Data.Pileup import naive_quick_pileup, naive_call_peaks
+from MACS3.Signal.Pileup import naive_quick_pileup, naive_call_peaks
 
 # ------------------------------------
 # Other modules
