+++ MACS3/Signal/SignalProcessing.pyx
