+++ MACS3/Signal/cPosValCalculation.c
