+++ MACS3/Signal/cPosValCalculation.h
