+++ MACS3/Signal/cPosValCalculation.pxd
