+++ MACS3/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["Commands","Utilities","IO","data"]
+__all__ = ["Commands","Utilities","IO","Signal"]
