+++ test/test_ScoreTrack.py
@@ -5,8 +5,8 @@ import io
 import unittest
 from numpy.testing import assert_equal,  assert_almost_equal, assert_array_equal
 
-from MACS3.Data.ScoreTrack import *
-from MACS3.Data.BedGraph import bedGraphTrackI
+from MACS3.Signal.ScoreTrack import *
+from MACS3.Signal.BedGraph import bedGraphTrackI
 
 class Test_TwoConditionScores(unittest.TestCase):
     def setUp(self):
